simple-rag/requirements.txt
beautifulsoup4==4.12.2
langchain==0.3.25
matplotlib==3.8.0
pre-commit==3.3.3
python-dotenv==1.0.0
ray==2.6.1
seaborn==0.12.2
sentence-transformers==2.2.2
slack_bolt==1.20.1
streamlit==1.24.1
structlog==23.3.0
typer==0.9.0
tiktoken==0.4.0
asyncpg==0.27.0
pgvector==0.2.4
psycopg[binary,pool]==3.1.11
psycopg2-binary==2.9.7
sqlalchemy[asyncio]==2.0.21
black==23.3.0
flake8==6.0.0
isort==5.12.0
python-multipart==0.0.6
pymupdf==1.22.5
furiosa-llm==2025.2.0
